Enrollment Templates management Menu Academic Management/Configuration/Enrollment templates to manage the enrollment templates.
Filtering subjects by study and only new subjects to avoid duplicates.
