+++ __manifest__.py
@@ -189,6 +189,8 @@
         'views/academic_management/enrollment_configuration/enrollment_authorization_search.xml',
         'views/academic_management/enrollment_configuration/enrollment_authorization_view.xml',
         'views/academic_management/enrollment_configuration/enrollment_authorization_form.xml',
+        'views/academic_management/enrollment_configuration/enrollment_template_form.xml',
+        'views/academic_management/enrollment_configuration/enrollment_template_view.xml',
         'views/academic_management/enrollment_configuration/menu.xml',
         
         'views/sales/product_view.xml',
