+++ models/enrollment/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 
-from . import (enrollment, enrollment_product_extension,enrollment_line_extension,authorization)
+from . import (enrollment_template, enrollment, enrollment_product_extension,enrollment_line_extension,authorization)
