+++ views/academic_management/enrollment_configuration/enrollment_template_view.xml
@@ -1,0 +1,42 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<odoo>
+    <data>
+        <record id="action_ems_enrollment_template" model="ir.actions.act_window">
+            <field name="name">Enrollment Templates</field>
+            <field name="res_model">sale.order.template</field>
+            <field name="view_mode">list,form</field>
+        </record>
+
+        <record id="view_sale_order_template_tree_inherit_ems" model="ir.ui.view">
+            <field name="name">sale.order.template.list.inherit.ems</field>
+            <field name="model">sale.order.template</field>
+            <field name="inherit_id" ref="sale_management.sale_order_template_view_tree"/>
+            <field name="arch" type="xml">
+                <xpath expr="//field[@name='name']" position="attributes">
+                    <attribute name="string">Enrollment Template</attribute>
+                </xpath>
+                <xpath expr="//field[@name='name']" position="after">
+                    <field name="ems_level_id"/>
+                    <field name="ems_study_id"/>
+                </xpath>
+            </field>
+        </record>
+
+        <record id="view_sale_order_template_search_inherit_ems" model="ir.ui.view">
+            <field name="name">sale.order.template.search.inherit.ems</field>
+            <field name="model">sale.order.template</field>
+            <field name="inherit_id" ref="sale_management.sale_order_template_view_search"/>
+            <field name="arch" type="xml">
+                <xpath expr="//search" position="inside">
+                    <field name="ems_level_id" string="Level"/>
+                    <field name="ems_study_id" string="Study"/>
+                    
+                    <group expand="0" string="Group By">
+                        <filter string="Level" name="groupby_level" context="{'group_by': 'ems_level_id'}"/>
+                        <filter string="Study" name="groupby_study" context="{'group_by': 'ems_study_id'}"/>
+                    </group>
+                </xpath>
+            </field>
+        </record>
+    </data>
+</odoo>
